requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
qrcode==8.0
Werkzeug==3.1.3
cloudinary==1.39.0
mysql-connector-python==8.0.33
python-dotenv==1.0.1